+++ eras/shelley/impl/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.19.0.0
 
-* Add `internalApplyTxWithValidation` to `ApplyTx`
+* Add `internalApplyTxWithValidation` and `internalReapplyValidatedTx` to `ApplyTx`
 * Change return type of `ruleApplyTxValidation` to `ValidatedTx`
 * Add:
   - `ValidatedTx`